Split Details.php
