+++ process/Command/GeoJSONCommand.php
@@ -3,7 +3,7 @@
 namespace App\Command;
 
 use App\Model\Config\Config;
-use App\Model\Details;
+use App\Model\Details\Details;
 use App\Model\GeoJSON\Feature;
 use App\Model\GeoJSON\FeatureCollection;
 use App\Model\GeoJSON\Geometry\Geometry;
