+++ process/Model/Details/LanguageValue.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Model\Details;
+
+class LanguageValue
+{
+    public string $language;
+    public string $value;
+}
